requirements.txt
flask
torch
torchvision
pillow
requests
gdown
